Rewrite redux store to TypeScript
