--- src/store/store.js
@@ -1,38 +1,0 @@
-import { compose, createStore, applyMiddleware } from 'redux';
-import { persistStore, persistReducer} from 'redux-persist';
-import storage from 'redux-persist/lib/storage';
-import logger from 'redux-logger';
-import createSagaMiddleware from '@redux-saga/core';
-
-
-import { rootReducer } from './root-reducer';
-import { rootSaga } from './root-saga';
-
-const persistConfig = {
-    key: 'root',
-    storage,
-    whitelist: ['cart']
-}
-
-const sagaMiddleWare = createSagaMiddleware();
-
-const persistedReducer = persistReducer(persistConfig, rootReducer);
-
-const middleWares = [
-    process.env.NODE_ENV !== 'production' && logger,
-    sagaMiddleWare,
-].filter(Boolean);
-
-
-const composedEnhancer = (
-    process.env.NODE_ENV !== 'production' && 
-    window && window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__
-    ) || compose;
-
-const composedEnhancers = composedEnhancer(applyMiddleware(...middleWares));
-
-export const store = createStore(persistedReducer, undefined, composedEnhancers);
-
-sagaMiddleWare.run(rootSaga);
-
-export const persistor = persistStore(store);
