+++ src/store/store.ts
@@ -1,0 +1,53 @@
+import { compose, createStore, applyMiddleware, Middleware } from "redux";
+import { persistStore, persistReducer, PersistConfig } from "redux-persist";
+import storage from "redux-persist/lib/storage";
+import logger from "redux-logger";
+import createSagaMiddleware from "@redux-saga/core";
+
+import { rootReducer } from "./root-reducer";
+import { rootSaga } from "./root-saga";
+
+export type RootState = ReturnType<typeof rootReducer>;
+
+declare global {
+  interface Window {
+    __REDUX_DEVTOOLS_EXTENSION_COMPOSE__?: typeof compose;
+  }
+}
+
+type ExtendedPersistConfig = PersistConfig<RootState> & {
+  whitelist: (keyof RootState)[];
+};
+
+const persistConfig: ExtendedPersistConfig = {
+  key: "root",
+  storage,
+  whitelist: ["cart"],
+};
+
+const sagaMiddleWare = createSagaMiddleware();
+
+const persistedReducer = persistReducer(persistConfig, rootReducer);
+
+const middleWares = [
+  process.env.NODE_ENV !== "production" && logger,
+  sagaMiddleWare,
+].filter((middleWare): middleWare is Middleware => Boolean(middleWare));
+
+const composedEnhancer =
+  (process.env.NODE_ENV !== "production" &&
+    window &&
+    window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__) ||
+  compose;
+
+const composedEnhancers = composedEnhancer(applyMiddleware(...middleWares));
+
+export const store = createStore(
+  persistedReducer,
+  undefined,
+  composedEnhancers
+);
+
+sagaMiddleWare.run(rootSaga);
+
+export const persistor = persistStore(store);
