+++ src/store/user/user.selector.ts
@@ -1,8 +1,9 @@
 import { createSelector } from "reselect";
 
 import { UserState } from "./user.reducer";
+import { RootState } from "../store";
 
-export const selectUserReducer = (state): UserState => state.user;
+export const selectUserReducer = (state: RootState): UserState => state.user;
 
 // Gets the currentUser from redux store
 export const selectCurrentUser = createSelector(
